Build CI sensitivity analysis B: stratification by 5FU chemo regimen
Bottom line: these results seem to indicate it's ok to combine the 5FU studies
and non-5FU studies for the Vis Mem domain, even though the pooled effect
sizes are less prominent among the 5FU studies. However, there is a big
difference between the 5FU and non-5FU studies along the Verb Mem domain. The
significant difference we see in Verb Mem in the main analysis is due to the
non-5FU studies.